Ref: common message lib for CV and CP (#117)
* Using serialization library

* minor simplification

* add missing message variants and fix docs

* port InvestmentManager to new serialization method

* fix rebase

* add support for more messages

* fix gas values

* port PoolManager to the new serialization method

* add MessageLib test file

* add identity tests

* MockCentrifugeChain partially updated

* port GasService to the new serialization method

* port Root to the new serialization method

* port Gateway to the new serialization method (except batches)

* introduce categories

* minor

* adapt tests

* adapt test and remove previous MessagesLib

* update gas

* minor rename

* minor reorder

* adapt gateway interface to receive chainId

* remove old comment

+++ src/vaults/CentrifugeRouter.sol
@@ -205,7 +205,7 @@ contract CentrifugeRouter is Auth, Multicall, ICentrifugeRouter {
     /// @inheritdoc ICentrifugeRouter
     function transferTrancheTokens(
         address vault,
-        uint64 chainId,
+        uint32 chainId,
         bytes32 recipient,
         uint128 amount,
         uint256 topUpAmount
@@ -221,7 +221,7 @@ contract CentrifugeRouter is Auth, Multicall, ICentrifugeRouter {
     /// @inheritdoc ICentrifugeRouter
     function transferTrancheTokens(
         address vault,
-        uint64 chainId,
+        uint32 chainId,
         address recipient,
         uint128 amount,
         uint256 topUpAmount
